+++ src/google/protobuf/compiler/java/lite/enum.cc
@@ -38,7 +38,7 @@ EnumLiteGenerator::EnumLiteGenerator(const EnumDescriptor* descriptor,
       immutable_api_(immutable_api),
       context_(context),
       name_resolver_(context->GetNameResolver()) {
-  for (size_t i = 0; i < descriptor_->value_count(); i++) {
+  for (int i = 0; i < descriptor_->value_count(); i++) {
     const EnumValueDescriptor* value = descriptor_->value(i);
     const EnumValueDescriptor* canonical_value =
         descriptor_->FindValueByNumber(value->number());
@@ -104,7 +104,7 @@ void EnumLiteGenerator::Generate(io::Printer* printer) {
     printer->Annotate("name", aliases_[i].value);
   }
 
-  for (size_t i = 0; i < descriptor_->value_count(); i++) {
+  for (int i = 0; i < descriptor_->value_count(); i++) {
     absl::flat_hash_map<absl::string_view, std::string> vars;
     vars["name"] = descriptor_->value(i)->name();
     vars["number"] = absl::StrCat(descriptor_->value(i)->number());
