+++ upb/io/string.h
@@ -102,7 +102,7 @@ UPB_INLINE bool upb_String_AppendFmtV(upb_String* s, const char* fmt,
   char* buf = (char*)malloc(capacity);
   bool out = false;
   for (;;) {
-    const int n = _upb_vsnprintf(buf, capacity, fmt, args);
+    const size_t n = _upb_vsnprintf(buf, capacity, fmt, args);
     if (n < 0) break;
     if (n < capacity) {
       out = upb_String_Append(s, buf, n);
